fix: solr readiness check added
